Begin with a grpc binding. Slice seems the lowest in the stack
I don't know how to make the finalizer to actually release the resources but the manual unref is working. This will be a long journey but I will be deeply excited if we get to the point to implement the stream protocol over buffer + slices instead of the copies we will initially do.

#[1 2 3] asGrpcSlice unref