Use >>#releaseData to avoid hidden leaks
Object>>#release is available and would be called in case a getter loses the type information. Use releaseData to make sure we release Grpc objects. Fix the initialization of the op array and we can now succesfully start a batch. It's time to try to make a real call and then figure out the initial completion queue integration.